Move MobileCartDialog into own file
